# backend/ops-admin/src/main/resources/openapi/opspanel-api.yaml
...
openapi: 3.0.3
info:
  title: OpsPanel Admin Backend API (Full Version)
  version: 1.0.0
  description: |
    Complete OpenAPI 3.0 specification for the OpsPanel Admin system.
    Includes:
      - Authentication
      - System management (user / role / menu / dept)
      - User-role relations
      - Role-menu relations
      - Login logs & operation logs
      - Health checks
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/system/user/list:
    get:
      tags: [User]
      summary: List users
      parameters:
        - name: pageNum
          in: query
          schema: { type: integer }
        - name: pageSize
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Paginated users
  /api/system/user:
    post:
      tags: [User]
      summary: Create user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateCmd'
      responses:
        '200': { description: Created }
    put:
      tags: [User]
      summary: Update user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateCmd'
      responses:
        '200': { description: Updated }
  /api/system/user/{id}:
    delete:
      tags: [User]
      summary: Delete user
      parameters:
        - name: id
          in: path
          schema: { type: integer }
      responses:
        '200': { description: Deleted }

  ############################
  # Role Management
  ############################
...